% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{Label}
\alias{Label}
\title{Google Compute Engine Instance Groups API Objects 
The Resource View API allows users to create and manage logical sets of Google Compute Engine instances.}
\usage{
Label(key = NULL, value = NULL)
}
\arguments{
\item{key}{Key of the label}

\item{value}{Value of the label}
}
\value{
Label object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:28:26
filename: /Users/mark/dev/R/autoGoogleAPI/googleresourceviewsv1beta2.auto/R/resourceviews_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Label Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The Label to be applied to the resource views.
}

