% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ListResourceResponseItem.endpoints}
\alias{ListResourceResponseItem.endpoints}
\title{ListResourceResponseItem.endpoints Object}
\usage{
ListResourceResponseItem.endpoints()
}
\value{
ListResourceResponseItem.endpoints object
}
\description{
ListResourceResponseItem.endpoints Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The list of service end points on the resource.
}
\seealso{
Other ListResourceResponseItem functions: \code{\link{ListResourceResponseItem}}
}
