% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ListResourceResponseItem}
\alias{ListResourceResponseItem}
\title{ListResourceResponseItem Object}
\usage{
ListResourceResponseItem(ListResourceResponseItem.endpoints = NULL,
  endpoints = NULL, resource = NULL)
}
\arguments{
\item{ListResourceResponseItem.endpoints}{The \link{ListResourceResponseItem.endpoints} object or list of objects}

\item{endpoints}{The list of service end points on the resource}

\item{resource}{The full URL of the resource}
}
\value{
ListResourceResponseItem object
}
\description{
ListResourceResponseItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The list response item that contains the resource and end points information.
}
\seealso{
Other ListResourceResponseItem functions: \code{\link{ListResourceResponseItem.endpoints}}
}
