% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{Operation.error.errors}
\alias{Operation.error.errors}
\title{Operation.error.errors Object}
\usage{
Operation.error.errors()
}
\value{
Operation.error.errors object
}
\description{
Operation.error.errors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Operation functions: \code{\link{Operation.error}},
  \code{\link{Operation.warnings.data}},
  \code{\link{Operation.warnings}}, \code{\link{Operation}}
}
