% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{Operation.warnings.data}
\alias{Operation.warnings.data}
\title{Operation.warnings.data Object}
\usage{
Operation.warnings.data()
}
\value{
Operation.warnings.data object
}
\description{
Operation.warnings.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Operation functions: \code{\link{Operation.error.errors}},
  \code{\link{Operation.error}},
  \code{\link{Operation.warnings}}, \code{\link{Operation}}
}

