% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_functions.R
\docType{package}
\name{resourceviews_googleAuthR}
\alias{resourceviews_googleAuthR}
\alias{resourceviews_googleAuthR-package}
\title{Google Compute Engine Instance Groups API
The Resource View API allows users to create and manage logical sets of Google Compute Engine instances.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:10:36
filename: /Users/mark/dev/R/autoGoogleAPI/googleresourceviewsv1beta2.auto/R/resourceviews_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
\item https://www.googleapis.com/auth/ndev.cloudman
\item https://www.googleapis.com/auth/ndev.cloudman.readonly
}
}
