% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ServiceEndpoint}
\alias{ServiceEndpoint}
\title{ServiceEndpoint Object}
\usage{
ServiceEndpoint(name = NULL, port = NULL)
}
\arguments{
\item{name}{The name of the service endpoint}

\item{port}{The port of the service endpoint}
}
\value{
ServiceEndpoint object
}
\description{
ServiceEndpoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The service endpoint that may be started in a VM.
}

