% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_functions.R
\name{zoneViews.getService}
\alias{zoneViews.getService}
\title{Get the service information of a resource view or a resource.}
\usage{
zoneViews.getService(project, zone, resourceView, resourceName = NULL)
}
\arguments{
\item{project}{The project name of the resource view}

\item{zone}{The zone name of the resource view}

\item{resourceView}{The name of the resource view}

\item{resourceName}{The name of the resource if user wants to get the service information of the resource}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/}{Google Documentation}
}

