% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_functions.R
\name{zoneViews.listResources}
\alias{zoneViews.listResources}
\title{List the resources of the resource view.}
\usage{
zoneViews.listResources(project, zone, resourceView, format = NULL,
  listState = NULL, maxResults = NULL, pageToken = NULL,
  serviceName = NULL)
}
\arguments{
\item{project}{The project name of the resource view}

\item{zone}{The zone name of the resource view}

\item{resourceView}{The name of the resource view}

\item{format}{The requested format of the return value}

\item{listState}{The state of the instance to list}

\item{maxResults}{Maximum count of results to be returned}

\item{pageToken}{Specifies a nextPageToken returned by a previous list request}

\item{serviceName}{The service name to return in the response}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
\item https://www.googleapis.com/auth/ndev.cloudman
\item https://www.googleapis.com/auth/ndev.cloudman.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly, https://www.googleapis.com/auth/ndev.cloudman, https://www.googleapis.com/auth/ndev.cloudman.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/}{Google Documentation}
}

