% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsGetServiceResponse}
\alias{ZoneViewsGetServiceResponse}
\title{ZoneViewsGetServiceResponse Object}
\usage{
ZoneViewsGetServiceResponse(endpoints = NULL, fingerprint = NULL)
}
\arguments{
\item{endpoints}{The service information}

\item{fingerprint}{The fingerprint of the service information}
}
\value{
ZoneViewsGetServiceResponse object
}
\description{
ZoneViewsGetServiceResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

