% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsList}
\alias{ZoneViewsList}
\title{ZoneViewsList Object}
\usage{
ZoneViewsList(items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{items}{The result that contains all resource views that meet the criteria}

\item{nextPageToken}{A token used for pagination}

\item{selfLink}{Server defined URL for this resource (output only)}
}
\value{
ZoneViewsList object
}
\description{
ZoneViewsList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a list request.
}

