% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsListResourcesResponse}
\alias{ZoneViewsListResourcesResponse}
\title{ZoneViewsListResourcesResponse Object}
\usage{
ZoneViewsListResourcesResponse(items = NULL, network = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{items}{The formatted JSON that is requested by the user}

\item{network}{The URL of a Compute Engine network to which the resources in the view belong}

\item{nextPageToken}{A token used for pagination}
}
\value{
ZoneViewsListResourcesResponse object
}
\description{
ZoneViewsListResourcesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response to a list resource request.
}

