% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourceviews_objects.R
\name{ZoneViewsSetServiceRequest}
\alias{ZoneViewsSetServiceRequest}
\title{ZoneViewsSetServiceRequest Object}
\usage{
ZoneViewsSetServiceRequest(endpoints = NULL, fingerprint = NULL,
  resourceName = NULL)
}
\arguments{
\item{endpoints}{The service information to be updated}

\item{fingerprint}{Fingerprint of the service information; a hash of the contents}

\item{resourceName}{The name of the resource if user wants to update the service information of the resource}
}
\value{
ZoneViewsSetServiceRequest object
}
\description{
ZoneViewsSetServiceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ZoneViewsSetServiceRequest functions: \code{\link{zoneViews.setService}}
}

